Rework bitstream syntax of opfl_refinement
Previously, inter compound mode is signaled using a symbol of 10
possible values (5 with opfl and 5 without opfl). Now, a bit
called use_optflow is transmitted first to indicate whether to use
opfl or not, then one out of five compound reference types is
signaled.

The impact in RD stats and speed is very small. Later, the
use_optflow symbol will be skipped under some conditions where opfl
is not allowed.

Change-Id: Ibc83bf3d3d9d6694e12e9e00f86c50d05a1f3637