Do not apply use_optflow in one sided compound
Disable optical flow MV refinement when reference blocks are both
from the past or both from the future. In such cases, the binary
symbol use_optflow is removed from the bitstream.

For random access, gain is increased by 0.04% with a 12% encoder
speedup.

For low delay, as previous, gain of opfl is neutral but the encoder
complexity increase in opfl is reduced from 25% to 5-10%.

Change-Id: I184dba43626efcaf17184b8499fea897121f5be7