Fix the decoder crash with CONFIG_SDP
The decoder crash happens when both sdp and intraBC are enabled for key frame.

In the previous code patches of SDP, variable tree_type exists in struct MB_MODE_INFO and struct MACROBLOCKD, which is easy to cause inconsistency. To solve the problem, variable tree_type from struct MB_MODE_INFO is removed in this code patch. And this variable is passed as the parameter into functions is_inter_block() and is_intrabc_block.

STATS_CHANGED

BUGS=aomedia:3052

Change-Id: I06c13c4be064ecb0d272f80af9d85a9c180cc83f