Enable passes=3 and pass=3.
Change oxcf->pass to use the enum properly.
Added AOM_RC_SECOND_PASS and AOM_RC_THIRD_PASS to the enum.

A new "passes" member is added to the encoder extra configs,
which can be controlled by the key & value API.

Currently the third pass is identical to the second pass.

Change-Id: I4f7280bc86b715c34c1001387e5284a85c4a8d7c