Cross-Component Sample Offset (CCSO)
In this work, Cross-Component Sample Offset (CCSO) approach is proposed to refine chroma component using luma component. The key component of CCSO is a non-linear offset mapping mechanism implemented as a look-up-table (LUT). The input of the mapping is the co-located reconstructed samples of luma component, and the output is offset values applied on chroma component.

Change-Id: I861dffcdc49c61fda831538274ccc5db3b387100