Simplify partition_search/pruning criterion
This is a refactoring commit that simplifies the criterion for
performing partition search and partition pruning. Many functional
signatures are changed to use only PartitionSearchState when possible.
Some helper functions for PartitionSearchState and PartitionBlkParams
are added to improve the readability of the code.

Furthermore, this commit also removes many redundant checks in the
condition. For example, to check whether partition_4 is allowed, the
codebase used to check for bsize >= BLOCK_8X8, do_rect_split, etc. But
those checks were already done during the assignment of
part4_search_allowed. In the refactored version, we simply check for
terminate_partition_search and part4_search_allowed. Similar
simplifications are done for other partition types as well with the
exception of normal rectangular split, as it would result in a
STATS_CHANGED due to interactions with is_active_edge.

Change-Id: I0c8f8a139eada79ab632088798a92e19fae61001