Simplify partition_search/pruning criterion
This is a refactoring commit that simplifies the criterion for
performing partition search and partition pruning. Functional
signatures are changed to use only PartitionSearchState when possible.
Some helper functions for PartitionSearchState and PartitionBlkParams
are added to improve the readability of the code.

Furthermore, this commit also removes redundant checks in search
conditions. For example, to check whether partition_4 is allowed, the
codebase used to check for bsize >= BLOCK_8X8, do_rect_split, etc. But
those checks were already done during the assignment of
part4_search_allowed. In the refactored version, we simply check for
terminate_partition_search and part4_search_allowed. Similar
simplifications are done for other partition types as well. The
exception of normal rectangular split, as it would result in a
stats_changed due to interactions with is_active_edge.

No stats_changed is expected, and the encoder is sped up by 0.02%.

Change-Id: I0c8f8a139eada79ab632088798a92e19fae61001