Normalize first-pass stats in the first pass.
Previously the stats are normalized in the second pass. This
could cause mismatches when the source of the first pass is not
of the same dimension with the source of the current pass.

Now the stats are all normalized in the first pass, and used
directly in the following passes.

Change-Id: Ia1f9a6196d15a76adce2ecb5563f70b586538ac4