Fix big quality drop with --allintra option
Resolve quality drop for speed 6 image coding (--allintra).

The --allintra setting (default for AVIF encoder) changes a number of
the speed quality trade off settings. Two of these for speed 6 cause
a very large drop in quality on many images.

The speed changes implemented when --allintra is selected at
speed 6 cause a drop in psnr of > 1db in some cases and >0.5db
in many others. Set against this the there is an improvement in encode
speed of over 50%.

This patch reverts two of the speed settings changes to the normal
speed 6 default. With this change most of the encode speed benefits
of --allintra remain but the drop in quality (psnr or perceptual) is
minimal.

Change-Id: I98e90b1e82ce0239ed581a9fb54aa58f896607b9