ERP: fix missing cdef strength assignment
When ERP is enabled, it is possible that a block is located in 2
CDEF units(64x64). In AV1, cdef strengths are parsed in regular
coding block scanning order, by first finding out the cdef unit
index and if nothing has been recorded for the corresponding 64x64
unit yet, read it from the bitstream. However, in ERP, if a block
is located in 2 CDEF units, it is necessary to find out the index
for the 2nd unit and parse the strength for it, otherwise cdef
strengths for some areas will be missing, resulting in recon
mismatches.

Change-Id: Iad2cc7a0850cf8c70f5d588edf307dee71ec5efd