REF_MV_BANK experiment to improve coding
Extend the number of motion vectors that can be accessed by
maintaining "banks" of them. A superblock at position (i, j) can access
row-bank "i" and column bank "j", and they are used when there are
open slots in the MV stack.

Original work by huisu@ on the research2 branch

Change-Id: I77a54cffff39ed4d4ca0c578d7a29e29e0efacd9