Collect transform coefficients frequency amplitude for analysis
Collect frequency amplitude from transform coefficients for analysis.

Intra prediction is applied to each 32x32 block.
The amplitude of transform coefficients is aggregated for the whole
frame.
These amplitudes can be used to compare the filtered image and
reconstructed image against the source image.

This cl is considered as an analysis tool.
A macro is set to turn off the analysis by default.

Change-Id: I0795fdddc9b5bf36eb4a16d4f65e5d455ac3cc97