Add two-pass input params and scaling
for three-pass encoding.

The idea of three-pass encoding is to use the bitstream from the
second pass to help with encoding the third pass. For speed
efficiency, a down-scaled video can be used for the second pass
for this purpose.

Two possible paths are designed for running three-pass encoding:

1. Running three-pass encoding using separate passes with libaom.

This CL adds the parameter "two-pass-output" to libaom, which
specifies the second-pass bitstream to be used by the third
pass. For example, one could run three-pass encoding by:

aomenc orig_input.y4m --passes=3 --pass=1 --fpf=fpf.txt \
       -o pass2.ivf
aomenc downscaled_input.y4m --passes=3 --pass=2 --fpf=fpf.txt \
       -o pass2.ivf
aomenc orig_input.y4m --passes=3 --pass=3 --fpf=fpf.txt \
       --two-pass-output=pass2.ivf -o output.ivf

This parameter is also exposed to the key & value API so one
could also use other programs (such as FFMpeg) to perform
three-pass encoding.

2. Running 3 passes altogether with aomenc.

In addition to "two-pass-output", this CL also adds the
parameters "two-pass-input", "two-pass-width" and
"two-pass-height" to aomenc. One could run three-pass encoding
by:

aomenc orig_input.y4m --passes=3 --two-pass-output=pass2.ivf \
       -o output.ivf

If two-pass-input is provided, the second pass will use it as the
input. If it is not provided, aomenc will rescale the frames in
the original input for the second pass.

The rescaled dimensions for the second pass is decided by the
following order:
- given by two-pass-width and two-pass-height
- parsed from the video file provided by two-pass-input
- downscale the original video file by 2 in both height and width

If two-pass-output is not provided, libaom will default to a
temporary file name "tmp_2pass_output_%d.ivf" where %d is the
stream index.

Note that currently this is a framework only and the third pass
does not do anything different from the second pass.

Bug=aomedia:3101

Change-Id: Ie0b6b8c57b9e0eefa1441805fbefdc41a7298b43