Disable DropFramesWithoutRecovery when altref is enabled
This test choses a subset of frames and forces them not to update
any reference buffers. This will not work with an altref or internal
altref, since a buffer will need to be updated in order to save the
frame until display time. A followup to unmark a frame as "droppable"
when it is coded out of order could be useful.

BUGS=aomedia:3090

Change-Id: I59922c205446a62d07153c290160282ab45f58b5