rtc: Fixes to the reference structure
Using golden as a reference was conditioned on
the nord_pickmode speeed feature, so speed 5 and 6
was not using it. Fix is to condition it on the rtc mode,
so all speeds for rtc mode may use golden as reference.

On a GF update (that is not key frame), refbuf_state should
be explicitly set to REFBUF_UPDATE, otherwise it will keep
the value of REFBUF_RESET on the GF udpate, which will also
refresh the moving ARF reference (which is used as reference).
This was not the intended behavior for rtc. The REFBUF states
were introduced in this CL, so this issue existed since then:
https://aomedia-review.googlesource.com/c/aom/+/117788

Also fix to avoid removing golden as reference in
get_ref_frame_flags(). The logic there was conditioned on
the nord_pickmode speeed feature, so speed 5 and 6 was not
using it. Fix is to condition it on the rtc mode, so all speeds
for rtc mode may use golden as reference.

avg. bdrate gains on speed 6 rtc_derf set: 4.3%
avg. bdrate gains on speed 7/8/9 rtc set: 1.8/1.78/0.7

Small speed slowdown ~1-2% observed on vga for speed >= 7.
Bigger slowdown for speed 6, ~10%, since GF is now enabled
as a reference.

Change-Id: I1c44004efb59c604f82231a882dc005cdb70999c