rtc: Fix to allow golden ref for speed < 7
Using golden as a reference was conditioned on
the nord_pickmode speeed feature, so speed 5 and 6
was not using it. Fix to condition it on the rtc mode,
so all speeds for rtc mode may use golden as reference.

avg. bdrate gain for speed 6:
~2.5% on rtc_derf/rtc set, several clips over 4%.
Slowdown ~5-8% on qvga and vga clips.

Change-Id: I1c44004efb59c604f82231a882dc005cdb70999c