rtc: Fixes to the reference structure
On a GF update (that is not key frame), refbuf_state should
be explicitly set to REFBUF_UPDATE, otherwise it will keep
the value of REFBUF_RESET on the GF udpate, which will also
refresh the moving ARF reference. This was not the intended
behavior for rtc. The REFBUF states were introduced in this CL:
https://aomedia-review.googlesource.com/c/aom/+/117788
so this issue has existed since then.

Also fix to avoid removing golden as reference in
get_ref_frame_flags(). The logic there was conditioned on
the nord_pickmode speeed feature, so for speed 5 and 6 the
golden could be removed as a reference. Fix is to condition it
on the rtc mode.

avg. bdrate gains on speed 7/8/9 rtc set: 1.8/1.78/0.7
Small speed slowdown ~1-2% observed on vga for speed >= 7.

For speed 6: bdrate gain ~4.3% but bigger slowdown, ~10%,
since GF is enabled as a reference with this change.
To avoid this, added speed feature to enable/disable golden ref
in pickmode, and disabled it for speed 5/6, so keep the same
behavior for now and avoid the ~10% slowdown.
bdrate gain for speed 6 with gf-ref speed feature off is 1.6%.

Change-Id: I1c44004efb59c604f82231a882dc005cdb70999c