Cleanup av1_rd_pick_palette_intra_sby
This CL unifies the handling of prune_palette_search_level for the
values 0 and 2. The separate handling for sf = 0 was introduced
when sf = 2 path performed the palette_size search in descending
order followed by ascending order. Now that the order is changed
in sf = 2 path, the sf = 0 path could be handled under a common
else part together with sf = 2 handling.

BUG=aomedia:3096

Change-Id: Id8eb3d9d9bc7f9feaa97090a28a17cf481753abb