Cleanup the handling of prune_palette_search_level = 0
This CL unifies the handling of the sf prune_palette_search_level
for the values 0 and 2. The separate handling for sf = 0 was
introduced when sf = 2 path performed the palette_size search in
descending order followed by ascending order. Now that the order
of evaluation is changed in sf = 2 path, the sf = 0 path could
also be handled under a common else part together with sf = 2
handling.

BUG=aomedia:3096

Change-Id: Id8eb3d9d9bc7f9feaa97090a28a17cf481753abb