rtc: Fix refbuf_state setting on GF update
On a GF update (that is not key frame), refbuf_state
should be explicitly set to REFBUF_UPDATE, otherwise
it will keep the value of REFBUF_RESET on the GF udpate,
which will also refresh the moving ARF reference (which
is used as reference for rtc). This was not the intended
behavior.

The REFBUF states were introduced in this CL, so this
issue existed since then.
https://aomedia-review.googlesource.com/c/aom/+/117788

avg. bdrate gains on speed 6 rtc_derf set: 4%
avg. bdrate gains on speed 7/8/9 rtc set: 1.8/1.78/0.7

Small speed slowdown ~1-2% observed on vga.

Change-Id: If871f6748b62440fabaea538417b837fd835493a