rtc: Set refbuf_state on GF update
On GF update that is not key frame, refbuf
should be explicitly set to REFBUF_UPDATE, otherwise
it will keep the value of REFBUF_RESET on the GF udpate,
which will also refresh the moving ARF reference (which
is used as reference for rtc).

bdrate gains on speed 6 rtc derf:
bdarate gains on speed 7/8/9 rtc set:

Small speed slowdown ~1% observed.

Change-Id: If871f6748b62440fabaea538417b837fd835493a