ThreePass: Read bitstream and derive GOP length.
Add functions to read and decode the two-pass bitstream for three pass
encoding.  A prototype to parse the GOP length is also added as a
proof of concept (will extend it to get the whole GOP structure in
future patches).

Currently only supports .ivf files for the input bitstream. This is to
avoid more potential concerns with complicated formats such as webm.

A few decoder control APIs are also added to retrieve relevant
information from the bitstream.

BUG=aomedia:3101

Change-Id: I87fdb0f332a5f2c4549f1dc2bd063ba3cba87ee8