Fix gm asserts in new reference signaling model
Occasionally, some of the original global parameters were being
set to default when the nrs global parameters were staying global,
causing failing asserts. This asserts the parameters in the two
frameworks are the same before the erroradvantage check, and then
copies the nrs parameters over to the regular framework parameters
to make sure they are the same.

Change-Id: Iaaa3260776633875c029477948a8fce58b8458ea