Take neighboring motion vectors into account.
Added a static inline method which looks at the up and left neighbors
(relative to current) of a motion vector and accounts for similarities
where the compression will be higher.

The scale_factor of 1.2 and mv_scale_factor of 5.0 are experimentally
chosen to yield a performance improvement. Against the baseline with the
CONFIG_BITRATE_ACCURACY flag off, these settings improve the overall
PSNR by 1.33% and the abs_rc_error by 8.70%.

Updated tpl_model_test accordingly and added more test cases.

BUG=aomedia:3045

Change-Id: I0b02ba786998e8c967d1ddd0d7f0fc8816904c06