ERP: copy cdef strength into all mi's
When SDP is on, every mi get its own MB_MODE_INFO allocated. In ERP,
cdef strengths are not always retrieved from the first mi of a coding
block, therefore we need to load the strength into all mi's of the
coding block in order to avoid a recon mismatch.

Change-Id: Idab2a67c9bfc838df52668e9481d96961d2530c9