Allintra: Reduce border_in_pixels for frame buffers
For allintra encoding mode, motion estimation is not performed and
the intraBC motion vectors are restricted within the tile boundary.
Hence, the border value used in the allocation of frame buffer is
reduced from 160 pixels to 64 pixels in case of allintra encoding
mode.

For AVIF image encode with speed = 9,

             HEAP Memory reduction(%)
Resolution   threads=1    threads=4
640x360        11.09        6.98
768x512         8.57        5.93
832x480         8.92        6.40
1280x720        6.33        5.03

For threads=4, an average encode time reduction of ~4.02% is
observed for 360p-720p resolutions.

HEAP memory reduction was measured using the following command.
$valgrind --tool=massif ./avifenc ...

Change-Id: Ibf690032b4c86a07f7751889de36b21e94544f1c