rtc: Fix condition on using golden for var_part
Condition should not be based on use_nonrd_pick_mode,
so speed 5 and 6 may use this, if golden is chosen
as a reference in speed features.

No change in behavior, as golden is not currently
used as reference. If golden reference is set/used,
then small bdrate gain on rtc_derf for speed 6: ~0.05%.

Change-Id: Ibfa7c62e1c7fbae254ca1c46cd8fd09d75c111d2