rtc: Allow for dual interpol filter in nonrd mode
This allows for interpol filter to be different in the
x and y directions, if the dual_filter speed feature is on.

Feature is disabled so no change in behavior.

If feature is enabled (sf->interp_sf.disable_dual_filter = 0)
not much gain was obtained: bdrate gain for speed 7/9 on
rtc set is small/negligible, most clips are neutral over rtc set.
And speed slowdown would be ~3-4%.

Change-Id: I6cb6571cc36c8cc0413b423297e1787259d3b1dc