Change  thresholds in intra_rd_variance_factor()
Reduce thresholds slightly for speeds 1-3 and disable for speeds >=4

For higher speeds there are fewer modes available in the search and
the risk of unpleasant artifacts is increased.

The change to the threshold for action based on speed means that
for speeds >= 4 the adjustment of the rd result will be disabled.

Change made as a response to:-
Issue 3138: Artifacts introduced by variance match weighting

Change-Id: Ib62e004a136f2342b2e0bfbbdaaed95025aec0d1