Assign q indices based on qstep_ratio.
Previously, we computed the highest level frame's qstep_ratio, then used
av1_q_mode_compute_gop_q_indices to determine the other frames' q
indices.

This change instead calculates the qstep_ratio for each frame in the GOP
and stores them in the vbr_rc_info struct. Then, each frame q index is
determined from the individual qstep_ratios.

We see a slight performance drop in PSNR and abs_rc_error in this CL.
This is because we use the unfiltered ARF when calculating the
qstep_ratios. This performance drop will be compensated for in a
following CL. The results for this CL are below.

              avg_psnr   ovr_psnr   abs_rc_error
previous CL   0.148%     -1.044%    -9.046%
this CL       0.384%     -0.588%    -7.979%

BUG=aomedia:3045

Change-Id: I6f40c5bff908e1898f25229337426fb4308c5b13