rtc: Adjust logic for gf refresh based on Q
After the QP is set just before encoding the frame,
check if we should force refresh or disable refresh of
golden frame, based on QP for current frame and past/avg_qp.
This will make the golden reference, which is used as second
longer term reference for 1 layer, somewhat better quality
and would generally improve the prediction.

~0.5% bdrate gain on rtc set for speed 9,
no speed change.

Change-Id: I4b2ea99e5f8e401d096e73be228bc3e7b2288a17