Update qstep_ratio_list after denoising ARF.
In the previous change, the qstep ratios were computed and used before
denoising the ARF frame. In this change, we try to improve performance
by recomputing the qstep ratio list after denoising the ARF.

We see a PSNR improvement in this change compared to the previous
change.

                         avg_psnr   ovr_psnr   abs_rc_error
before qstep_ratio_list  0.148%     -1.044%    -9.046%
previous CL              0.384%     -0.588%    -7.979%
this CL                  -0.305%    -1.434%    -4.283%

BUG=aomedia:3045

Change-Id: Ibaedbd613d88d46218179f596e879b710da5826c