Update qstep_ratio_list after denoising ARF.
In the previous change
(https://aomedia-review.googlesource.com/c/aom/+/145745), the qstep ratios
were computed and used before denoising the ARF frame. In this change, we
try to improve performance by recomputing the qstep ratio list after
denoising the ARF.

Note that we see ~0.4% BDRate improvement but 5% abs_rc_error setback.
The 0.4% BDRate improvement shows that TPL stats is helpful to determine
q indices in a GOP.

The 5% abs_rc_error setback is probably because we are still estimating
the base layer q index before filtering ARF.

We will work on improving abs_rc_error after this CL is checked in.

The results are measured with the lowres dataset with limit=17.

                         avg_psnr   ovr_psnr   abs_rc_error
before qstep_ratio_list  0.148%     -1.044%    -9.046%
previous CL              0.384%     -0.588%    -7.979%
this CL                  -0.305%    -1.434%    -4.283%

BUG=aomedia:3045

Change-Id: Ibaedbd613d88d46218179f596e879b710da5826c