rtc: Some adjustments for temporal layers
cyclic refresh: increase the percent_refresh,
account for larger interval for key frames,
and add condition to disable for spatial layers
whose base is key.

Use the larger motion search also on TL1 frames,
for 3 temporal layers.

And scale the target size on the very first (key)frame,
to compensate for lower framerate on TL0.

Change-Id: I07aba41d0eebe7090a6e21c9704cc7a4e3b991dc