rtc: Fix to lm=3 pattern in svc sample encoder
For the 3TL pattern with multiple references (lm=3):
fix the ALTREF ref_idx so that the ALTREF trails
the enhancement layer frames (in current code it
was only trailing on the TL0 frames), so TL2/TL1
frames can also use the ALTREF as a reference.

Also enable compound prediction for (LAST, ALTREF)
for speed >= 7 for pattern lm = 3.

bdrate improvemnt of ~1-2% on  rtc_set for speed 7,
3TL with lm=3.

Change-Id: Ia5def436e7c7c6a27cb74dfe33d231f862022cf0