rtc: Cleanup/fixes for aq-mode=3
Remove the last_coded_q, as this was not
currently being updated and hence used.
This removes an allocation for 1 layer and
multi-layer.

The aq-mode=3 segment feature for RTC is only used for
blocks >= 8X8, so skipping over 4x4 block can gives
small speedup (~1-1.5%) at speed 10, vga.

But there is small bdrate loss (~0.2%), since
for more efficient coding of the seg map, the
cur_frame->seg_map needs to set at 4x4, along with the
function av1_cyclic_reset_segment_skip().
This was mentioned here:
https://aomedia-review.googlesource.com/c/aom/+/136503

Change-Id: I15efe2909aab108678fa15f043fd1ab00f3eec8c