Add tune=image_perceptual_quality
Similar to tune=ssim, this mode adjusts rd-multipliers of blocks as small
as 16x16 to redistribute the bits used for encoding based on human
perceptual quality.

The model used in this mode is based on user rating data set (same as
deltaq-mode=4). It is recommended to use this mode with deltaq-mode=4
and avif/image compression.

Change-Id: Ibea59ba2c3c1091d8a0dea895a68ca4154eed62d