Allintra: Pick an adaptive step size for normalization
The normalization factor in deltaq-mode=3 is important to determine
proper qindex for each superblock.

The current step size in the calculation of the normalization factor
is the superblock size (64x64).

However, we could imagine contents that the image characteristics
inside a superblock vary significantly. In such cases, we might
want to use a smaller block size as the step size in the computation
of the normalization factor.

In this change, we propose to determine the step size adaptively by
comparing the source mean and variance of the original and the
sub block size.

If we found the statistics indicate there is a strong local variance,
we split the original block size. A majority vote of all superblocks
finally decides whether to use a smaller block size or not.
The process goes recursively, until 16x16.

Change-Id: I57feb6f5d2ffc0ac4db82fce60cb6c7848e64ec1