Only compute tpl stats when gf_frame_index = 0
Previously in the first GOP, we will compute tpl stats
at key frame with gf_frame_index = 1 and ARF frame with
gf_frame_index = 0. In this CL, we merged the to runs
of TPL stats computation.

We see mostly neutral results except for midres there is 0.1~0.2% drop
and for hdres there is 0.1% gain

q mode --lag-in-frames=35 --limit=150

         (negative is better)
         avg_psnr ovr_psnr ssim
ugc360p   0.038  0.040  -0.018
lowres   -0.045 -0.029  -0.080
midres    0.104  0.149   0.213
hdres    -0.102 -0.094  -0.124

Note that, this result is measured a few weeks ago.
I will measure the performance again and update the results
in the commit message before submitting the CL

BUG=aomedia:3144

STATS_CHANGED

Change-Id: I3216d86c54f721019110b58f9c7a1befbdeea9b5