rtc: Update conditions for selective cdf_update mode
This only  affects cdf-update-mode = 2 (selective mode)
for real-time mode.

Move the logic for disable/enable to a separate function,
and account for some use cases (layers, resize, scene-change).

The update can have significant reduction in frame size
(~20% in some cases) on key frames or on scene changes,
so explicity add those conditions, along with resized frames.

Also keep update on for a few frames before the GF update,
as this can help with the quality of the GF reference.

To avoid bdrate loss, keep forcing udpate every ~10 frames for now.

For speed 10 on rtc set with cdf-update-mode=2:
bdrate loss ~0.8%. Small speedup observed on vga clip (as updates
are less than doing it every other frames) of ~2%.

Change-Id: I2286558d72d6bcacd80de47272b1892ae2a39256