write_features_to_file: check fopen() result
This is a follow up to:
dadc6023d ExternalPartitionTestDfsAPI: skip check w/WRITE_FEATURE_TO_FILE=0

Return if the file open fails to avoid a crash. In the case of
WRITE_FEATURE_TO_FILE=0 the test will pass as expected. With
WRITE_FEATURE_TO_FILE=1 this would cause test failure as the comparison
will be performed which is also expected.

Bug: b/191266788
Change-Id: I7fa52b5599533843f6d17c22970583e664a217be