write_features_to_file: fix early return logic
This is a follow up to:
dadc6023d ExternalPartitionTestDfsAPI: skip check w/WRITE_FEATURE_TO_FILE=0

previously if WRITE_FEATURE_TO_FILE was false and is_test_mode was
true the file would still be written which could cause a crash if the
file open operation failed. this change corrects the check and adds a
check on the fopen return.

Bug: b/191266788
Change-Id: I7fa52b5599533843f6d17c22970583e664a217be