rtc: Adjust gf length for speed >= 9
At real time speed 9 and 10, due to the deep optimizations, using
avg_frame_low_motion to represent high or low motion isn't accurate
anymore. To compensate for that, we lower the gf length in general.
Later, if we find better stats to allow the encoder to decide the
gf length adaptively, we could revert this change.

STATS_CHANGED

Change-Id: I561ad8914efd8a0c5231c40871d32dca6c7ad5fe