rtc: Alllow source_sad selectively for speed >= 10
Enable source_sad speed feature for speed >= 10,
but only selectively based on the current frame's
avg_source_sad (running average source_sad) and QP.

The source_sad is used to condition the increase
in partition thresholds and usage of modeled_nonrd_cost.
Enabling it locally gets ~4% bdrate gain on desktop_ex2,
and ~2% on several other clips, with avg bdrate gain ~0.5%
on rtc set. This is with --cdf-update-mode=2.

Speed slowdown measured on vga clips is ~2%.

Change-Id: I5aae2fa60f14182934da66b3d4bc7c761709f6bd