Fix printing of frame size in console output
The existing implementation of this is not very useful as the frame
size is printed and then immediately deleted, i.e.
Pass 2/2 {etc...} [ETA  xx:xx:xx]  {time passes while frame is encoded}  24523F \033[K

This means that the frame size (24523F here) generally does not
appear in the console, or if it does, only for a very brief flash.

This commit changes the behaviour so the frame size is printed alongside
the other statistics so that is visible. i.e.
Pass 2/2 {etc...} [ETA  xx:xx:xx] 24523F  {time passes while frame is encoded}  \033[K

Change-Id: Iae30d335767527d5aab3af88b02e57df8dd7099c