Fix printing of frame size in console output
The existing implementation of this is not very useful as the frame
size is printed and then immediately deleted, and so generally does
not appear in the console, or if it does, only for a very brief flash.

This commit changes the behaviour so the frame size is printed alongside
the other statistics so that is visible.

Change-Id: Iae30d335767527d5aab3af88b02e57df8dd7099c