Increase default value of cpu_used
Currently the default is cpu-used 0 which is largely impractical
for everyday use, and only offers reasonably small efficiency
improvements over faster modes while being significantly slower.

This commit sets the default to cpu-used 4 which seems to be a
reasonable middle ground, although of course it can be discussed
whether this should be tweaked one way or the other.

Change-Id: I3a9bf624995a4abca3e4e69f587cf90d8f97019e