Allintra: Adjust DCM dynamic range by MSCN contrast
The max quantizer used in the DCM model (deltaq-model=3) is bounded
by that value such that transform coefficients are not quantized to
zero.

The min quantizer however is previously based on a heuristic, i.e.,
the half of the frame quantizer.

Here we use the "mean subtracted contrast normalized (MSCN)"
coefficients to estimate the structure information of each
super block.

Define the contrast as the quotient between the abs sum of MSCN
coefficients of two super blocks. We then use the max contrast to
estimate the min quantizer.

This effectively improves the dynamic range estimation, and leads
to a better bits distribution and finally improved perceptual quality.

Change-Id: I45986acf1e6c9f922e1bb0e111ea0ea3a676dcbc