rtc: Speed feature to switch to half_pel at block level
Add speed feature, set at speed >= 10, to allow for
switching to half_pel for newmv at coding block level.
Switch is based on the size of the integer motion vector,
along with bsize and avg_frame_low_motion.

avg. bdrate loss ~0.6 on rtc_set, max is 1.7 for dark720p
speedup ~2% on some vga clips. This is for speed 10,
with cdf-update-mode=2.

Change-Id: I38201e2a9d5ac8336045b2b58168686b541d2933