Only call av1_temporal_filter when needed
Move the logics of checking second arf out of av1_temporal_filter.

Move the showable_frame logics out of av1_temporal_filter.

Before, we assumed av1_temporal_filter will be called when
it is about to be coded. But that's not the case anymore.
Therefore, we need to move those logics out.

BUG=aomedia:3144

Change-Id: I4ed5ded65a1d559abc66638b0a4ee0c7948cac6b