Add partition pruning strategy for thirdpass.
In general, as twopass encoding typically uses a small
resolution, it is usually safe to assume that the block size
used in the thirdpass is smaller or equal to the (upscaled)
block size from the second pass.

This patch utilizes such observation and applies a few
pruning rules for partition search in the third pass.

Overall on youtube_ugc set, with minimal speed
influence (~0.7% slower), BD-rate was reduced by 0.324% and
0.491% for ovr_PSNR and SSIM.

BUG=aomedia:3101

STATS_CHANGED

Change-Id: I44b6d3a2bbdddaccc79b0c18adf6114999392845