Allintra: Introduce adapt_top_model_rd_count_using_neighbors sf
The sf top_intra_model_count_allowed indicates the number of top
model rds to be stored during intra mode decision and these stored
model rds are used to prune the luma intra modes. This CL
introduces aggressive pruning of modes by reducing the number of
candidates used to store the model rd and by adapting the index of
the candidate for pruning based on neighbor blocks and quantizer
information. This change is applicable for speed >= 6.

For AVIF still-image encode,

          Instruction Count   BD-Rate Loss(%)
cpu-used     Reduction(%)     psnr       ssim
    6           5.481         0.0222    -0.0281
    7           4.170         0.2591     0.2361
    8           2.437         0.1108     0.1157
    9           0.000         0.0000     0.0000

STATS_CHANGED

Change-Id: Ib1bf8cae2f78c640c387050e6a2d4dfb988e742e