Allintra: Delay top-right sync for row-mt
When row-mt is enabled, a thread waits for the encoding of
the top-right superblock to be complete before encoding the
current superblock. This wait time is reduced in allintra
mode, by delaying the sync wait until the data from top-right
superblock is needed. This change is applicable for
resolutions < 4k in speed >= 9 as cost update frequencies
are set to COST_UPD_OFF for these resolutions.

Change-Id: I3cdeb7ad58df976788cc6bd058184dfc65b63593