svc test for error resilience: dropping mid-sequence
This test checks for continued successful decoding
when a set of enhancement layers are dropped within
middle of sequence (instead of dropping all TL1/TL2, as in
the other existing tests). This test more closely captures
the real application case where any set/sequence of TL1/TL2
may be dropped (due to packet loss) and then decoding resumes
on a TL0 frame.

Succesful decoding with no mismatch is guaranteed in this test
if the error resilience is enabled (1) on enhancement layers,
but can remain off (0) on base TL0.

Change-Id: I32d747945a007df484fbf779736b983643a3b420