Allintra: Turn off several intra coding tools automatically
Add a command line option to automatically determine whether to
turn off several intra coding tools to get faster encoding speed.
By default the feature is off.

This mode is effective only for all intra mode when "--deltaq-mode=3"
is true, i.e., the perceptual quality mode.

When the flag is set, the encoder analyzes the encoding condition
and the reconstructed image quality as compared to the source in
the preprocessing stange performed in deltaq-mode.

If it is in the high quality range, and the reconstructed image
quality is good enough, we turn off several intra coding tools,
based on the observation that in the high quality range,
the perceptual quality difference is less likely perceivable even
without a few intra coding tools that have an impact on the
compression efficiency.

With this change, in high quality range, the encoding speed is
up to 80% faster than the baseline on some images.
40% to 50% faster is usually seen.
There is no impact in low quality range.

Change-Id: I21566f14fbc7b226632bbcf48d955f8108cbbdc2