rtc: Fix prediction on keyframes for spatial layers
On superframes whose base is key: fix the reference
predicton structure so that the SL layer (for SL > 0)
only does the inter-layer prediction from the SL-1 layer.
This requires fixes to the sample encoder and in the
nonrd_pickmode.

In the current code SL2 was either predicting only from
SL0 (for ksvc), or it was predicting from both SL0 & SL1
(in non-ksvc mode), which was not the intended behavior.
So this fix improves the key frame quality for ksvc and
reduces the encode_time for non-ksvc.

Change-Id: Ifde750c741b53b42e4f27d493945036fb7e0e334