Remove nonrd condition in var based partitioning
Prior to this commit, the condition was added to prevent a regression.
The regression was due to an incorrect segment id, which led to the
wrong q index being used.

This commit fixes the issue by replacing
av1_set_offsets_without_segment_id with av1_set_offsets.

Performance:
 SPD_SET | TESTSET | AVG_PSNR | OVR_PSNR |  SSIM   |  SPD
---------+---------+----------+----------+---------+-------
    5    |   rtc   | -0.617%  | -0.487%  | -0.923% | +0.9%
    5    |rtc_derf | -0.256%  | -0.243%  | -0.178% | -2.0%
---------+---------+----------+----------+---------+-------
    6    |   rtc   | -1.283%  | -1.147%  | -1.593% | -0.3%
    6    |rtc_derf | -0.127%  | -0.199%  | -0.120% | -1.6%
---------+---------+----------+----------+---------+-------

STATS_CHANGED

Change-Id: I643cde296344bfa68e0f9c655bfed0d85cd7e404