rtc-svc: Fix to avoid removing references
get_ref_frame_flags() removes references which
are found to be duplicates to higher priority ones.
Add condition that in order for a reference to be
removed relative to the higher priority one, the latter
reference is actually used as reference. This is relevant
for SVC where the user selects which references are used,
and in some case LAST (highest priority one) is not actually
used.

Also in non_pickmode: only allow for pruning/removing
golden and altref if LAST is used as a reference
(i.e., ref_frame_flags & AOM_LAST_FLAG).

This prevents the assert failure in nonrd_pickmode that at
least one of the references (last, golden, altref) is available.
The assert is currently re-enabled in this CL for svc mode.

Modified svc datarate unittest to trigger the assert failure
without this fix in get_ref_frame_flags().

Change-Id: I8f77ad87b280e9f80693c9a1686b6fef1164bc3e