AV1 RT: Implement merging blocks with same mode
if the smallest blocks in partition have same mode, ref frame and no
residual we can merge them into a larger block of the same mode.
This gives ~0.4% BDRate gain, but because we ahve to do extra
initialization and cannot re-use prediction we need to spend ~3% more CPU
cycles. Not worth it...

Change-Id: Ib836b3b30103de79fc3e410f185507365af45f7d