rtc: Fixes and some adjustments to cyclic refresh
Correct the update to the actual_seg#_blocks for
skip blocks. Also avoid using the rate/dist threshold
at low-resoln or for rd_pickmode, as they need to be
better tuned/adjusted for that case.
And adjust some parameters mainly for low-resoln case.

bdrate gains on rtc_derf for speed 6/7/9: ~1.1/1.4/0.27%
neutral/small loss on rtc set for speed 7/9: ~0.27/0.5.

Change-Id: I4a5f82ace14f16798efbdc18c897e88de07c98b8