Resolve TODO: Inactive_zone_cols
Place holder but for now no plans to implement.

The inactive zones (rows and columns) act to optimize behavior
in two pass encoding content where there is a dead zone added
to pad the real image to a different aspect ratio.

This is comment in the case of vertical padding or "letterboxing"
but less so for padding portrait format content horizontally,
which tends to only apply to slide shows of portrait stills.

The code here is a place holder mirroring the behavior for rows,
but there are no plans to implement this at present.

Change-Id: Idc9f3ba3b7e6a3927555fe2e14a62b0d54eb1227