Add a flag to use quantization matrices during RD search.
The flag "dist-metric", when set to "qm-psnr", will weight the error in
transform space by the quantization matrix; this should be a
psychovisual improvement. Not setting the flag, or setting it to "psnr",
will not change the current behaviour.

Note that this flag is not tested (and may not work) with 64-point
transforms, as those transforms discard some coefficients before
quantization.

BUG=aomedia:3209

Change-Id: Ied5a1d9bb127263de19becf7302b37219f70a324