rtc: Fix to boundary check in set_vt_partitioning
This reducess rectangular blocks on the boundary,
in particular for 180x320 clips and gives some
speedup and bdrate gain for those clips.

For speed 9 on rtc_derf: most of the 180p clips
get about ~2-3% speedup with ~0.5-1% bdrate gain.
Negligible change on the other clips.
Average speedup: 1.1%

No change for 480p and above (where sb_size == 128).

Change-Id: Ia55aee79c68549e74771a6cb6481d147b2c58b17