Avoid the unnecessary computation of CfL rd cost
During the evaluation of CfL mode, the rd cost for each chroma plane
is computed after the relevant alpha parameter is determined. However
if alpha parameters of both the chroma planes are zeroes, then the
computed rd cost is discarded as it is not a valid scenario. In this
CL, alpha parameters of both the chroma planes are determined first
followed by the computation of rd cost. Hence the rd cost calculation
is avoided when alpha parameters of both the chroma planes are zeroes.
This CL is applicable only when the sf cfl_search_range is set to 1.

For AVIF still-image encode,

             Encode Time
cpu-used     Reduction(%)
    6           1.161
    7           1.773
    8           0.961

Change-Id: Ia0539df15842d33f349cda68758113f57dcc5839