Avoid the unnecessary computation of CfL rd cost
During the evaluation of CfL mode, the rd cost for each plane is
computed after the relavant alpha parameter is determined. However
if alpha parameters of both the planes are zeroes, then the computed
rd cost is discarded as it is not a valid scenario. In this CL, alpha
parameters of both the planes are determined first followed by the
computation of rd cost for CfL mode. Hence the rd cost calculation is
avoided when alpha parameters of both the planes are zeroes. This CL
is applicable only when the sf cfl_search_range is set to 1.

For AVIF still-image encode,

             Encode Time
cpu-used     Reduction(%)
    6           0.978
    7           0.921
    8           1.312

Change-Id: Ia0539df15842d33f349cda68758113f57dcc5839