In-place temporal filtering in RTC encoder
Implemented a lightweight temporal filter in real-time encoder.
Used the past frame(s) to filter the current frame. Boosted the
coding efficiency and also reduced the noise. The in-place
approach didn't need an extra frame buffer allocation and
reduced the memory access time, and, as a result, the encoder
time wasn't increased.

STATS_CHANGED

Change-Id: I1807bd0cafe49f86d228693d93689e67b38d2cd4