Don't pass the TileInfo struct by value
The TileInfo struct consists of six 'int' fields. It is not a small
struct, so we should not pass it by value. Also, we sometimes copy a
TileInfo struct to a local variable so that we can refer to the struct
by a shorter name. That kind of copy should be replaced by taking a
const pointer to the struct.

Change-Id: I94cc493b6c923359bb55623e5bdb4d74e6012f61