Replace AOM_EXT_PART_ABI_VERSION with old value
In the definition of AOM_ENCODER_ABI_VERSION, replace
AOM_EXT_PART_ABI_VERSION with its value (3) in libaom v3.2.0, so as to
not break the ABI version check in aom_codec_enc_init_ver() when an
application compiled against libaom v3.2.0 passes the old value of
AOM_ENCODER_ABI_VERSION to aom_codec_enc_init_ver().

Mark the external partition API as experimental. This was announced in
the libaom v3.2.0 release notes, but I neglected to add a note to the
source code.

Bug: 3236
Change-Id: Iefbe11ebce94c1ae7ca1e5f05f434eb71824b8fa