Allintra: Fix the initialization of intra_sb_rdmult_modifier
It was observed that the value of rdmult was turning out to be zero
in non-RD path after the following CL:
https://aomedia-review.googlesource.com/c/aom/+/153361
This was happening because the initialization of
intra_sb_rdmult_modifier was not being done when encode_nonrd_sb()
was invoked. This patch fixes the same by making the initialization
of intra_sb_rdmult_modifier common for both encode_rd_sb() and
encode_nonrd_sb().

STATS_CHANGED for non-RD encode

Change-Id: Ia3f0493d2c74ee4425f83214fe28e7f8db0c5817