rtc: Enable sse_early_term for sp10 and avoid skip test
The compuations for setting this_early_term in
model_skip_for_sb_y_large() can be avoided when sse_early_term
is enabled and calculate_rd 0. This is bitexact for speed <= 9
and gets ~0.5% speedup.

For speed 10: enable sse_early_term with same level as speed 9:
speedup is ~1% with bdrate loss:

max loss is 1.6% but that clip (desktop360p) is still ~25%
better bdrate than vp9 reference.

Change-Id: I954ecb9ffce021c348b978a6811402b90e627b6a