Simplify and fix log_sub_block_var()
Remove read and write back of the sub block variance.

The function log_sub_block_var() was attempting to read
variance values from a data structure, if valid, and write back to
the same structure. However, this function is only currently
called for the ALL_INTRA case where the structure will not
be populated. Furthermore, the write back was not correct
and in experiments where calls to this function were made
for all key frames (as opposed to just single still images)
it caused some non deterministic behavior and borg crashes.

Use of saved variance values can be revisited in the future if
necessary.

Change-Id: If45d72298a8868614017d46e2fad6efa5dbe0c3b