RT: use lower qindex after 15 frames.
Following the parent CL's logic, we should boost the quality of frames
to a high level as early as possible so that they can serve as good
predictors for following frames and thus increase compression
efficiency.

But the constraint of WebRTC states that within 0.5 sec, we don't
want to see a large overshoot. The key frame is apparently an overshot,
therefore we delay 15 frames to do the boost.

Change-Id: I5cd6bb13330b9625b6c2b0d0c47cb2af6a846ab5