rtc-screen: Early exit in variance part for zeromv
For screen content: allow for early exit out of the
variance partitioning (for static blocks on base segment)
and force zeromv-last skip mode in nonrd_pickmode.

Enabled only for speed 10 for now.
This speeds up encoder on rtc_screen clips, ranging
from ~3 to 10% speedup.
bdrate gain of ~1%, max loss is 0.8%

Change-Id: Id821cd98eab218a19d624e0c1d76880c5e23f237