Allintra: Avoid the allocation of sms_tree
The structure 'sms_tree' is used to store the simple motion search
data for partition pruning in inter frames. The same is not
required during allintra encoding. Hence, the relevant memory
allocations and initializations are avoided for allintra encoding
mode.

For AVIF image encode with speed = 9,

             HEAP Memory reduction(%)
Resolution   threads=1    threads=4
640x360        0.47         1.47
768x512        1.16         3.65
832x480        0.32         1.11
1280x720       0.63         2.15

For threads=4, an average encode time reduction of ~2.32% is
observed for 360p-720p resolutions.

HEAP memory reduction was measured using the following command.
$valgrind --tool=massif ./avifenc ...

Change-Id: Idd6dff396f0abd505167c769c1e3ef3c37b2a16a