rtc: Add additonal split condition in variance part
Add additonal condition to split from 32x32 to 16x16
in variance condition: based on the sub16x16 variance
and the source content metrics, and for the case where
the force_large_partition_blocks is enabled (speed >= 9).

This is needed to reduce visual artifacts where big blocks
(32x32) are being used in moving area (dragging artifact).
For example, in the desktopqvga2 clip.

~2% bdrate gain speed 9 rtc_derf, max gain is 7.5%.
~3-4% average slowdown measured in encoder-fps at 100k, over
the rtc_derf testset.

Change-Id: I1f7095b211d20d55d0e3ca6b473d5a7a14fd9795