rtc: Fix to blk_slip for screen
The blk_skip was not being updated properly
intra modes or palette mode, and for tx_size = 4x4.
This mainly affects screen, which uses tx_4x4 for intra.
For video in nonrd the tx_size >= 8X8.

bdrate neutral/negligible for screen and rtc set.
Speedup ~2-6% on slide changes for speed 10.

Change-Id: Ic14aadee457c9b14e956caf1b7ea482831f6501d