rtc: Fix to blk_skip update for screen
The blk_skip was not being updated properly
palette or intra modes, which may use tx_size = 4x4.
This mainly affects screen, which uses 4x4 for intra.
For inter modes in nonrd tx_size >= 8X8 and so the
blk_skip is kept at 8x8 level.

bdrate neutral/negligible for screen and rtc set.
Speedup ~2-6% on slide changes for speed 10.

Change-Id: Idd2e53cf4b87d1362b2514c5df1da58e76573917