rtc: Fix to blk_skip update for screen
The blk_skip was not being updated properly for
palette or intra modes, which may use tx_size = 4x4.
This mainly affects screen, which uses 4x4 for intra.
For inter modes in nonrd tx_size >= 8X8 and so the
blk_skip is kept at 8x8 level.

The only effect this has for video mode is for 32x32 blocks
that pick DC mode, because these would use 16x16 tx_size,
and the blk_skip (if set) was not being passed to ctx.
But this rarely happens, so change in rtc sets are negligible.

bdrate neutral for screen, speedup ~1-5% on some slide changes
for speed 10.

Change-Id: Idd2e53cf4b87d1362b2514c5df1da58e76573917