Move the allocation of MB_MODE_INFO_EXT_FRAME buffer to encode stage
The MB_MODE_INFO_EXT_FRAME buffer is used to store the mode
information during the encode stage. This information is consumed
later during the bitstream preparation stage. This buffer is
allocated during the init stage by assuming a minimum partition
size of BLOCK_4X4. This CL moves the allocation of this buffer to
the encode stage so that the allocation is done on the basis of
the sf default_min_partition_size.

For AVIF image encode with speed = 9,

             HEAP Memory reduction(%)
Resolution   threads=1    threads=4
640x360        11.25         9.12
768x512        12.19         9.91
832x480        12.82        11.06
1280x720       14.31        12.76

For threads=4, an average encode time reduction of ~5.38% is
observed for 360p-720p resolutions.

HEAP memory reduction was measured using the following command.
$valgrind --tool=massif ./avifenc ...

Change-Id: If09a2a36f60022ed163b516cc0e81788c1ff7ccb