Allocate dv_costs when required
The structure 'dv_costs' is used to hold the cost of signalling
the dv in the bitstream in case of intraBC mode. This patch
allocates 'dv_costs' only when intraBC mode needs to be evaluated.
As 'dv_costs' is not used during first pass, the code related
to its allocation is cleaned up from fp_prepare_enc_workers().

For AVIF image encode with speed = 9,

             HEAP Memory reduction(%)
Resolution   threads=1    threads=4
640x360        3.25         2.64
768x512        2.06         2.28
832x480        2.14         1.84
1280x720       1.03         0.92

HEAP memory reduction was measured using the following command.
$valgrind --tool=massif ./avifenc ...

Change-Id: If40abe0bde968edc067b56143e9320078f57dad9