Add a test of MockRateControlQMode
Normally you wouldn't bother to test a mock, but this is the first
usage of gmock here, so it's a way to verify that it builds.

This reqiured making gmock build correctly in all configutaions.

Also cleaned up the aom_gtest target so that its implementation
headers are private, its interface headers are public (but only
to transitive dependencies), and made the tests that use it
link to it explicitly.

Change-Id: I7c0ee70f43f3e410fc4232d8d14e1496877cda28