Allintra: Avoid global table initializations
The global tables related to wedge mask and active Q Index are not
required for allintra, AOM_Q encode. This CL avoids the
initialization of these tables in this case.

For AVIF still-image encode, an average encode time reduction
of ~2.45% is observed on average across different resolutions with
speed=9 and 4 threads.

Change-Id: Ie2a1df372b8c0b9a0d67ad339974285a658ef183