Simplify the condition for freeing dv_costs
In multi-threaded encoding, the dv_costs pointer value is copied
from the master thread for all the worker threads. When dv cost
update is enabled, this pointer needs to be reallocated on need
basis. Hence, the pointer is reset to NULL in such cases prior to
the conditional allocation such that the conditional expression
for calling aom_free() is simplified for the same. Also added
comments highlighting the need to keep track of the conditional
expressions being used for alloc/free operations specific to
worker threads.

BUG=aomedia:3268

Change-Id: Ibfbb430d7036f8bccb693d3b00cc77d41622e55d