rtc: Set sb_size to 64 for 720p and below.
In i--sb-size=0 (dynamic/default), and for
real-time mode (or for spatial layers),
keep the sb_size to 64 for:
MIN(width, height) <= 720 (instead of 480).

This has good bdrate gains on 720p clips,
and small speedup on average over 720p.

bdrate gains for speed 10 range from
~1 to ~7% gains, several clips 5% gains.
Average is 1.5%.

average speedup over 720p clips in rtc set,
for speed 10: ~1%.

Change-Id: Iae640b582dd3edc7dbb6997f79faf519351d6825