Allintra: Introduce prune_luma_odd_delta_angles_in_intra sf
The odd delta angles of directional luma intra modes are pruned
using the RD-costs of neighboring delta angles. To facilitate this
optimization, the order of evaluation of delta angles is changed
when the sf prune_luma_odd_delta_angles_in_intra is enabled.
This sf is enabled for speed >= 6.

For AVIF still-image encode,

             Encode Time      BD-Rate Loss(%)
cpu-used     Reduction(%)     psnr       ssim
    6           2.601         0.2491     0.2657
    7           2.389         0.1043     0.1236
    8           2.032         0.0669     0.0721

STATS_CHANGED

Change-Id: Idf92adbc8f58ffbfa2706ebb6a520e0a749880e6