rtc: Add test for error_resilience for (3SL, 3TL)
Add test for svc-rtc for 3 spatial and 3 temporal
layers, where a set of temporal enhancement layers
are dropped mid-sequence, and then decoding is started
again at next (SL0, TL0) frame. Verfy that successful
decoding can be done with error_resilience off/0.

Successful decoding with lost TL2/TL1 frames is possible
due to the nature of the temporal prediction structures
used for svc-rtc (LAST is always temporal refereance and
refers to lower temporal layer), and the setting of the
primary_ref_frame when error_resilience is off:
https://aomedia-review.googlesource.com/c/aom/+/157583

Change-Id: I1b336d989a232bbc601bec663704160ed35ab876