Set frame border size based on sb_size
Currently, the frame border size is hard-coded to 128+32=160 pixels.
But in the RTC use case, often the superblock size is set to 64X64, so
the border padding only needs to be 64+32=96 pixels. This commit makes
the encoder to set the border size based on the current sb_size.

This change gives about 10% memory reduction and some minor speedup.

Performance:

Change-Id: I3628d02bc02e201dc4e13a5dd3efd5d94c24eb5c