Set frame border size based on sb_size
Currently, the frame border size is hard-coded to 128+32=160 pixels.
But in the RTC use case, often the superblock size is set to 64X64, so
the border padding only needs to be 64+32=96 pixels. This commit makes
the encoder to set the border size based on the current sb_size.

This change gives about 10% memory reduction and some minor speedup.

Performance:
 | SPD_SET | TEST_SET | ENC_TIME  |
 |---------|----------|-----------|
 |    6    | rtc_derf |  -0.205%  |
 |    6    |   rtc    |  -0.043%  |
 |---------|----------|-----------|
 |    7    | rtc_derf |  -0.478%  |
 |    7    |   rtc    |  -0.133%  |
 |---------|----------|-----------|
 |    8    | rtc_derf |  -0.641%  |
 |    8    |   rtc    |  -0.174%  |
 |---------|----------|-----------|
 |    9    | rtc_derf |  -0.996%  |
 |    9    |   rtc    |  -0.248%  |
 |---------|----------|-----------|
 |   10    | rtc_derf |  -1.127%  |
 |   10    |   rtc    |  -0.296%  |
 |---------|----------|-----------|

Change-Id: I3628d02bc02e201dc4e13a5dd3efd5d94c24eb5c