pickrst,update_[ab]_sep_sym: refine clamping
based on a rough analysis with 12-bit input, assuming Y[] values < 2048
and accumulating 128 squared Y[] values, H[] may use (signed) 26 bits.
a/b need to be clamped to (signed) 28 bits to allow the descaled result
of H * [ab] to fit in (signed) 64 bits. this may need further
adjustment if more values are accumulated or the range of Y[] goes to
12 bits.

Bug: b/231820849
Change-Id: I114b171be8057c634064b009e66fd01a15ced751