svc-rtc: Fix to multi-ref pattern in test
For both the unittest and sample encoder:
when testing the multi-ref (3SL, 3TL) pattern, don't
allow altref as an additional reference for superframes
whose base is key. Since that will allow for prediciton
from SL2 to the base SL0 (1/16 resolution), instead of
only using the previous SL1 (1/4 resolution, which is the
intended behavior).

Also initialize svc->skip_mvsearch_altref to 0 in restore_layer_context();

Change-Id: I2a0cf64052af9c7a1dc431d3dd94f50fd57438f6