rtc: Use pre-analysis sb source_ad for aq-mode=3
For aq-mode=3 (cyclic refresh): use the superblock
source_sad, computed in the scene detection for rtc,
into the cyclic_refresh for better targeting of the
segment for delta-qp.

Avoid setting superblock to segment if that sb has
high block source_sad, since it would likely get reset
after the pickmode. And allow it to be set if cr->map=1
(previous block was intra or high motion mode) but is
not low source_sad. The idea is to get the segment set
faster to areas where it will be useful for cleanup.

bdrate gains for rtc_derf/rtc at speed 9: 0.65/0.55,
mclips have some bdrate gain, a few 2%.
Negligible encoder-fps change observed.

Change-Id: I73d8ab9842b1e58d4d2b59b030e093fff2e33784