rtc: Use pre-analysis sb source_ad for aq-mode=3
For aq-mode=3 (cyclic refresh): use the superblock
source_sad, computed in the scene detection for rtc,
into the cyclic_refresh for better targeting of the
segment for delta-qp.

Avoid setting superblock to segment if that sb has
high block source_sad, since it would likely get reset
after the pickmode. And allow it to be set if cr->map=1
(previous block was intra or high motion mode) but is
not low source_sad. The idea is to get the segment set
faster to areas where it will be useful for cleanup.

Only for non-screen mode for now.
And only affects 720p for speed 9 and 10, where
the block_sad is saved in the scene detection.

bdrate gains for rtc_derf/rtc at speed 10 (avg/ovr/ssim):
rtc: -0.24/-0.21/-0.43.
Most clips have ~0.5-1% bdrate gain, worst loss is 0.3%.

Negligible encoder-fps change observed.

Change-Id: I73d8ab9842b1e58d4d2b59b030e093fff2e33784