RESIZE_MODE: Fix incorrect strides being used for motion search
Currently during motion search, the encoder always assumes that the ref
frame's stride is the same as that of source frame. But this assumption
breaks when resize mode/superres feature is turned on, leading to
unexpected motion search result and invalid memory access.

This commit actively checks for the ref stride with what's stored in
MotionVectorSearchParams::search_site_config, and creates a new
search_site_config if there is a mismatch.

BUG=aomedia:3283

Change-Id: Ia99a88326bf716027b5a652ec519fb13cfa0a345