rtc-screen: Fix to skipping newmv in nonrd
For spatially flat blocks with non-zero motion
inter-modes with zero motion are skipped. But
we should also skip NEWMV, if the motion vector
is (0, 0). This condition for skip newmv needs to
be added after search_newmv.
Without this fix an assert will be triggered in
encode_mv if nearest or near are also (0, 0).

Add test with many scene cuts and high motion,
this triggers the assert in the issue without this fix.

Bug: aomedia:3289
Change-Id: Ib52c67d22ccf66b39a43c449b8edc91204917017