Implement sad<w>xh_neon functions using UDOT instruction
Add an alternative implementation of sad128xh_neon, sad64xh_neon,
sad32xh_neon and sad16xh_neon using ADB, UDOT instead of ADB, UADALP.
UDOT is higher throughput and lower latency than UADALP on modern Arm
CPUs and accumulating into 32-bit elements removes the need for an
extended reduction at the end of the loop.

The existing implementation is retained for CPUs that do not
implement the Armv8.4-A UDOT instruction. The availability of the
UDOT instruction is indicated by the feature macro
__ARM_FEATURE_DOTPROD.

Change-Id: Ic0f2f99d76b480da7277c734400168d6ace3d462