Allintra: Introduce aggressive tx_domain_dist_level
This CL enables the usage of transform domain distortion during the
winner mode evaluation stage for allintra speed>=6. The same is
achieved by introducing an aggressive level for the speed feature
tx_domain_dist_level. The documentation for this speed feature is
also updated to be synonymous with its implementation.

For AVIF still-image encode,

             Encode Time      BD-Rate Loss(%)
cpu-used     Reduction(%)     psnr       ssim
    6           4.480         0.0863     0.0730
    7           2.848         0.0714     0.0696
    8           3.120         0.0295     0.0145

STATS_CHANGED

Change-Id: I5832e301a74c5671bb3240078b8d719e833b5277